voice-recorder-backend/requirements.txt
openai-whisper
pymongo
python-dotenv
spacy